chore: ampl warning message and governance page text fix
